Enhance form handling and introduce new field types
- Added new field types: `multiselect`, `image`, and `collection`.
- Introduced components for `MultiSelectInput`, `ImageInput`, and `CollectionInput`.
- Enhanced `FormService` to manage `FormArray` controls for collections.
- Refactored `BaseFormComponent` to delegate form handling to `FormService`.
- Updated `form-renderer` to support new input types and conditional visibility.
- Added a 12-column grid system for responsive form layouts.
- Improved `BlobHelper` to check blob existence before deletion.
- Simplified route initialization with new providers and `resolve` logic.

+++ backend/src/lrembecki.infrastructure/Helpers/BlobHelper.cs
@@ -47,7 +47,10 @@ internal class BlobHelper(BlobServiceClient serviceClient) : IBlobHelper
         var containerClient = serviceClient.GetBlobContainerClient(container);
         var blobClient = containerClient.GetBlobClient(blobName);
 
-        await blobClient.DeleteAsync(DeleteSnapshotsOption.IncludeSnapshots, cancellationToken: cancellationToken);
+        if ((await blobClient.ExistsAsync(cancellationToken)).Value)
+        {
+            await blobClient.DeleteAsync(DeleteSnapshotsOption.IncludeSnapshots, cancellationToken: cancellationToken);
+        }
     }
 
     private static string CalculateContentType(string filename)
